requirements.txt
celery==5.2.1
channels==2.3.0
channels-redis==3.3.1
Django==3.2.9
django-celery==3.1.17
django-celery-results==2.2.0
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
gevent==21.8.0
nest-asyncio==1.5.1
psycopg2-binary==2.9.2
pytz==2021.3
redis==4.0.2
